tutor/requirements.txt
# Google AI/ML and Cloud Libraries
# Pinned to ensure the exact versions that work with the agent kit
a2a-sdk==0.3.10
google-adk==1.18.0
google-genai==1.48.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.5.0
google-cloud-logging==3.12.1

# Core Web Framework Stack (Pinned for conflict resolution and stability)
fastapi==0.116.1
starlette==0.47.2
uvicorn[standard]==0.35.0
